estilização de paginação e imagens para produto

+++ business/business/ItemPedido.cs
@@ -1,3 +1,6 @@
+using System;
+
+
 namespace business.business
 {
       public class ItemPedido : BaseModel
@@ -16,9 +19,9 @@ namespace business.business
         }
 
         public int Quantidade { get; set; }
-        public int ProdutoId { get; set; }
+        public Int64 ProdutoId { get; set; }
         public virtual Produto Produto { get; set; }
-        public int PedidoId { get; set; }
+        public Int64 PedidoId { get; set; }
         public virtual Pedido Pedido { get; set; }
         public decimal PrecoUnitario { get; set; }
